✨ refactor(document): split FormDetail into Detail and DetailWrapper components
