+++ app/components/document/form/DetailWrapper.vue
@@ -1,0 +1,13 @@
+<template>
+  <ShadcnCard>
+    <ShadcnCardHeader>
+      <ShadcnCardTitle>Details</ShadcnCardTitle>
+      <ShadcnCardDescription>
+        The transaction details.
+      </ShadcnCardDescription>
+    </ShadcnCardHeader>
+    <ShadcnCardContent>
+      <slot />
+    </ShadcnCardContent>
+  </ShadcnCard>
+</template>
